chore: always run scripts against this repo (#11)
* chore: run postinstall to apply previous changes

* deps: tap@15.0.10

* chore: always run scripts against this repo

Since this repo applies all of the changes to itself, if the scripts
are not run manually then there can be unexpected git churn when running
other commands. This change adds behavior and a test to fail if any of
the scripts will cause changes locally.

With this new behavior test coverage was no longer 100%, so this also
increases coverage and refactors some tests.