fix: use conventional commits from release-please for changelog (#183)
`release-please` already fetches the commits and parses them into
conventional commit objects, so we are able to reuse most of that
instead of fetching it from GitHub again. This also adds tests for the
changelog output.

This also removes the workspace-deps plugin in favor of extending the
builtin node-workspace plugin. This fixes the issue of workspaces
sometimes not getting the correct tag name and changelog title if they
were only bumped as part of a workspace dep.