feat: make template-oss postinstall commit a breaking change for majors (#180)
When we ship things like `engines` changes we will release a semver
major update to template-oss, and the resulting dependabot PRs will get
a `feat!` commit to indicate that they will all need major releases as
well. This doesn't assume anything about the contents of the breaking
change, but will instead set the correct conventional commit type and
fail the PR so that it can be followed up manually.