feat: add overwrite false property to added files
By default template-oss allows for written files to be configured on a
per-repo basis. This is helpful for different repos to create their own
templated files and apply those.

With this change, a repo can now set overwrite: false to a templated
file and have those updates be applied after the default template-oss
changes are made.