fix: pin release-please version (#388)
- fix: pin release-please version

There is a tradeoff with our usage of release-please. We are using some
`release-please` classes that we import directly from files like `const
{ ManifestPlugin } = require('release-please/build/src/plugin.js')`. Due
to how load bearing our release process is for us, I think it is best to
always pin to an exact version and use dependabot PRs to review all
`release-please` changes.

- chore: update release please mocks and snapshots

This PR also updates mocks and snapshots. I was not expecting this to
need to be updated but they failed in #385. It's possible that the mock
recording or the usage of these mocks inside release-please is
non-deterministic. I will keep an eye on these tests and refactor in the
future if they turn out to be flaky.