fix: pass releases to publish check (#444)
If `publish: true` is not set, then the release intergration workflow
falls back to running `npm view $SPEC` which will error if the package
has not been published. This tags the CODEOWNERS in the release issue so
we know something went wrong.

This doesn't have much use now that we are moving everything to
`publish: true` but it would have caught a missing publish last week
that I missed on `@npmcli/fs`:
https://github.com/npm/fs/actions/runs/8946801618/job/24577963034#step:7:4

The `RELEASES` env var was only being set for the publish step, not the
check if published step.